find pb and heartbeat

+++ client/00DF_runtime.go
@@ -34,7 +34,7 @@ func f00DE8010strcat(dst []uint8, src string) {
 	dst = append(dst, src...)
 }
 
-func f00DE8100memset(buf []uint8, value uint8, size uint32) uint32 {
+func f00DE8100memset(buf []uint8, value uint8, size int) uint32 {
 	if size == 0 {
 		return 0
 	}
@@ -364,7 +364,7 @@ func main() {
 }
 
 // 大范围清零操作
-func f00DFC986(buf []uint8, size uint32) {
+func f00DFC986(buf []uint8, size int) {
 	size >>= 7
 	for size != 0 {
 		// 它使用了8次 movdqa xmm指令
@@ -378,7 +378,7 @@ func f00DFC986(buf []uint8, size uint32) {
 
 // setzero， 16字节对齐
 // 某种意义上说，把迭代改为递归也可以有效防止饥饿调度问题
-func f00DFC9DD(buf []uint8, value uint8, size uint32) uint32 {
+func f00DFC9DD(buf []uint8, value uint8, size int) uint32 {
 	// 假设eax是0x0018,DF85，然后edx是0x3FF
 	// cdq          ;edx:eax组合成为64位
 	// mov edi,eax
@@ -394,7 +394,7 @@ func f00DFC9DD(buf []uint8, value uint8, size uint32) uint32 {
 
 	// offset := 16 - uint32(&buf[0])%16
 	// offset := 16 - uint32(unsafe.Pointer(&buf))%16
-	offset := 16 - (*[3]uint)(unsafe.Pointer(&buf))[0]%16
+	offset := 16 - (*[3]int)(unsafe.Pointer(&buf))[0]%16
 	if offset == 0 {
 		s := size & 0x7F
 		if s != size {
@@ -414,7 +414,7 @@ func f00DFC9DD(buf []uint8, value uint8, size uint32) uint32 {
 		buf[i] = 0
 	}
 
-	size -= uint32(offset)
+	size -= offset
 	f00DFC9DD(buf[offset:], 0, size)
 	return 0
 }
